requirements.txt
certifi==2020.12.5
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-PyMongo==2.3.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.3
Werkzeug==1.0.1
wincertstore==0.2
psycopg2>=2.8.6
pandas==1.1.3